fwk: Move InvalidDataException to fwk/Exceptions
